+++ code/ComplexPlane.h
@@ -10,9 +10,9 @@ using namespace sf;
 using namespace std;
 
 const unsigned int MAX_ITER = 64;
-const float BASE_WIDTH = 4.0;
-const float BASE_HEIGHT = 4.0;
-const float BASE_ZOOM = 0.5;
+const double BASE_WIDTH = 4.0;
+const double BASE_HEIGHT = 4.0;
+const double BASE_ZOOM = 0.5;
 
 enum class State {
     CALCULATING,
@@ -28,7 +28,7 @@ private:
 	Vector2f m_plane_center;
 	Vector2f m_plane_size;
 	int m_zoomCount;
-	float m_aspectRatio;
+	double m_aspectRatio;
 
 	int countIterations(Vector2f coord);
 	void iterationsToRGB(size_t count, Uint8& r, Uint8& g, Uint8& b);
